agent-backend/requirements.txt
streamlit==1.31.1
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
elasticsearch==8.11.0
folium==0.15.1
geopy==2.4.1
openai-agents
rich
fastapi
uvicorn
langfuse
pydantic-ai[logfire]
nest_asyncio
datasets
pymongo
bcrypt
python-jose==3.3.0
zep-cloud